chore: split entrypoint of checkdep
